Refactor job functions
