+++ setup.py
@@ -58,15 +58,15 @@ setuptools.setup(
         ],
         "biologic": [
             "biologic-kbio",
-        ]
+        ],
     },
     entry_points={
         "console_scripts": [
             "tomato=tomato:run_tomato",
             "ketchup=tomato:run_ketchup",
-            "tomato-job=tomato.drivers:tomato_job",
+            "tomato-job=tomato.daemon.job:tomato_job",
             "tomato-driver=tomato.daemon.driver:tomato_driver",
-            "tomato-daemon=tomato.daemon:run_daemon",
+            "tomato-daemon=tomato.daemon:tomato_daemon",
         ]
     },
 )
