+++ src/tomato/daemon/job.py
@@ -11,19 +11,24 @@
     be propagated via the :class:`tomato.daemon.cmd` set of functions.
 
 """
+
 import os
 import subprocess
 import logging
 import json
 import time
+import argparse
+from importlib import metadata
 from datetime import datetime, timezone
 from pathlib import Path
 from threading import currentThread
+from multiprocessing import Process
 
 import zmq
 import psutil
 
-from tomato.models import Pipeline, Daemon
+from tomato.daemon.io import merge_netcdfs, data_to_pickle
+from tomato.models import Pipeline, Daemon, Component, Device, Driver
 
 logger = logging.getLogger(__name__)
 
@@ -238,3 +243,256 @@ def manager(port: int, timeout: int = 500):
         action_queued_jobs(daemon, matched_pips, req)
         time.sleep(timeout / 1e3)
     logger.info("instructed to quit")
+
+
+def tomato_job() -> None:
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--version",
+        action="version",
+        version=f'%(prog)s version {metadata.version("tomato")}',
+    )
+    parser.add_argument(
+        "--port",
+        help="Port on which tomato-daemon is listening.",
+        default=1234,
+        type=int,
+    )
+    parser.add_argument(
+        "--timeout",
+        help="Timeout [ms] for driver actions.",
+        default=1000,
+        type=int,
+    )
+    parser.add_argument(
+        "jobfile",
+        type=Path,
+        help="Path to a ketchup-processed payload json file.",
+    )
+    args = parser.parse_args()
+
+    with args.jobfile.open() as infile:
+        jsdata = json.load(infile)
+    payload = jsdata["payload"]
+    ready = payload["tomato"].get("unlock_when_done", False)
+    pip = jsdata["pipeline"]["name"]
+    jobid = jsdata["job"]["id"]
+    jobpath = Path(jsdata["job"]["path"]).resolve()
+
+    logfile = jobpath / f"job-{jobid}.log"
+    logging.basicConfig(
+        level=logging.DEBUG,
+        format="%(asctime)s - %(levelname)8s - %(name)-30s - %(message)s",
+        handlers=[logging.FileHandler(logfile, mode="a")],
+    )
+    logger = logging.getLogger(__name__)
+
+    tomato = payload.get("tomato", {})
+    verbosity = tomato.get("verbosity", "INFO")
+    loglevel = logging._checkLevel(verbosity)
+    logger.debug("setting logger verbosity to '%s'", verbosity)
+    logger.setLevel(loglevel)
+
+    if psutil.WINDOWS:
+        pid = os.getppid()
+    elif psutil.POSIX:
+        pid = os.getpid()
+
+    logger.debug(f"assigning job {jobid} with pid {pid} into pipeline {pip!r}")
+    context = zmq.Context()
+    req = context.socket(zmq.REQ)
+    req.connect(f"tcp://127.0.0.1:{args.port}")
+    poller = zmq.Poller()
+    poller.register(req, zmq.POLLIN)
+    params = dict(pid=pid, status="r", executed_at=str(datetime.now(timezone.utc)))
+    req.send_pyobj(dict(cmd="job", id=jobid, params=params))
+    events = dict(poller.poll(args.timeout))
+    if req in events:
+        req.recv_pyobj()
+    else:
+        logger.warning(f"could not contact tomato-daemon in {args.timeout/1000} s")
+
+    output = tomato["output"]
+    prefix = f"results.{jobid}" if output["prefix"] is None else output["prefix"]
+    outpath = Path(output["path"])
+    snappath = outpath / f"snapshot.{jobid}.nc"
+    logger.debug(f"output folder is {outpath}")
+    if outpath.exists():
+        assert outpath.is_dir()
+    else:
+        logger.debug("path does not exist, creating")
+        os.makedirs(outpath)
+
+    logger.info("handing off to 'job_worker'")
+    logger.info("==============================")
+    ret = job_worker(context, args.port, payload, pip, jobpath, snappath)
+    logger.info("==============================")
+
+    merge_netcdfs(jobpath, outpath / f"{prefix}.nc")
+
+    if ret is None:
+        logger.info("job finished successfully, attempting to set status to 'c'")
+        params = dict(status="c", completed_at=str(datetime.now(timezone.utc)))
+        req.send_pyobj(dict(cmd="job", id=jobid, params=params))
+        events = dict(poller.poll(args.timeout))
+        if req not in events:
+            logger.warning(f"could not contact tomato-daemon in {args.timeout/1000} s")
+            req.setsockopt(zmq.LINGER, 0)
+            req.close()
+            poller.unregister(req)
+            req = context.socket(zmq.REQ)
+            req.connect(f"tcp://127.0.0.1:{args.port}")
+        else:
+            ret = req.recv_pyobj()
+            logger.debug(f"{ret=}")
+            if ret.success is False:
+                logger.error("could not set job status for unknown reason")
+                return 1
+    else:
+        logger.info("job was terminated, status should be 'cd'")
+        logger.info("handing off to 'driver_reset'")
+        logger.info("==============================")
+        # driver_reset(pip)
+        logger.info("==============================")
+        ready = False
+    logger.info(f"resetting pipeline {pip!r}")
+    params = dict(jobid=None, ready=ready, name=pip)
+    req.send_pyobj(dict(cmd="pipeline", params=params))
+    events = dict(poller.poll(args.timeout))
+    if req in events:
+        ret = req.recv_pyobj()
+        logger.debug(f"{ret=}")
+        if not ret.success:
+            logger.error(f"could not reset pipeline {pip!r}")
+            return 1
+    else:
+        logger.error(f"could not contact tomato-daemon in {args.timeout/1000} s")
+        return 1
+    logger.info("exiting tomato-job")
+
+
+def job_process(
+    tasks: list,
+    component: Component,
+    device: Device,
+    driver: Driver,
+    jobpath: Path,
+):
+    sender = f"{__name__}.job_process"
+    logger = logging.getLogger(sender)
+    logger.debug(f"in job process of {component.role!r}")
+
+    context = zmq.Context()
+    req = context.socket(zmq.REQ)
+    req.connect(f"tcp://127.0.0.1:{driver.port}")
+    logger.debug(f"job process of {component.role!r} connected to tomato-daemon")
+
+    kwargs = dict(address=component.address, channel=component.channel)
+
+    datapath = jobpath / f"{component.role}.pkl"
+    logger.debug("distributing tasks:")
+    for task in tasks:
+        logger.debug(f"{task=}")
+        while True:
+            req.send_pyobj(dict(cmd="task_status", params={**kwargs}))
+            ret = req.recv_pyobj()
+            logger.debug(f"{ret=}")
+            if ret.success and ret.msg == "ready":
+                break
+
+        req.send_pyobj(dict(cmd="task_start", params={**task, **kwargs}))
+        ret = req.recv_pyobj()
+        logger.debug(f"{ret=}")
+
+        t0 = time.perf_counter()
+        while True:
+            tN = time.perf_counter()
+            if tN - t0 > device.pollrate:
+                req.send_pyobj(dict(cmd="task_data", params={**kwargs}))
+                ret = req.recv_pyobj()
+                if ret.success:
+                    data_to_pickle(ret.data, datapath)
+                t0 += device.pollrate
+            req.send_pyobj(dict(cmd="task_status", params={**kwargs}))
+            ret = req.recv_pyobj()
+            logger.debug(f"{ret=}")
+            if ret.success and ret.msg == "ready":
+                break
+            time.sleep(device.pollrate / 10)
+        req.send_pyobj(dict(cmd="task_data", params={**kwargs}))
+        ret = req.recv_pyobj()
+        if ret.success:
+            data_to_pickle(ret.data, datapath)
+
+
+def job_worker(
+    context: zmq.Context,
+    port: int,
+    payload: dict,
+    pipname: str,
+    jobpath: Path,
+    snappath: Path,
+) -> None:
+    sender = f"{__name__}.job_worker"
+    logger = logging.getLogger(sender)
+
+    req = context.socket(zmq.REQ)
+    req.connect(f"tcp://127.0.0.1:{port}")
+
+    while True:
+        req.send_pyobj(dict(cmd="status", with_data=True, sender=sender))
+        daemon = req.recv_pyobj().data
+        if all([drv.port is not None for drv in daemon.drvs.values()]):
+            break
+        else:
+            logger.debug("not all tomato-drivers have a port, waiting")
+            time.sleep(1)
+
+    pipeline = daemon.pips[pipname]
+    logger.debug(f"{pipeline=}")
+
+    # collate steps by role
+    plan = {}
+    for step in payload["method"]:
+        if step["device"] not in plan:
+            plan[step["device"]] = []
+        task = {k: v for k, v in step.items()}
+        del task["device"]
+        task["task"] = task.pop("technique")
+        plan[step["device"]].append(task)
+    logger.debug(f"{plan=}")
+
+    # distribute plan into threads
+    processes = {}
+    for role, tasks in plan.items():
+        component = pipeline.devs[role]
+        logger.debug(f"{component=}")
+        device = daemon.devs[component.name]
+        logger.debug(f"{device=}")
+        driver = daemon.drvs[device.driver]
+        logger.debug(f"{driver=}")
+        processes[role] = Process(
+            target=job_process,
+            args=(tasks, component, device, driver, jobpath),
+            name="job-process",
+        )
+        processes[role].start()
+
+    # wait until threads join or we're killed
+    snapshot = payload["tomato"].get("snapshot", None)
+    t0 = time.perf_counter()
+    while True:
+        tN = time.perf_counter()
+        if snapshot is not None and tN - t0 > snapshot["frequency"]:
+            logger.debug("creating snapshot")
+            merge_netcdfs(jobpath, snappath)
+            t0 += snapshot["frequency"]
+        joined = [proc.is_alive() is False for proc in processes.values()]
+        if all(joined):
+            break
+        else:
+            time.sleep(1)
+    logger.debug(f"{[proc.exitcode for proc in processes.values()]}")
+    for proc in processes.values():
+        if proc.exitcode != 0:
+            return proc.exitcode
